pyproject.toml
[tool.poetry]
name = "ansible-vault-filter"
version = "0.1.0"
description = "Hooks for my ansible configuration."
authors = ["m09 <142691+m09@users.noreply.github.com>"]
license = "Apache-2.0"
readme = "README.md"
homepage = "https://github.com/m09/ansible"

[tool.poetry.dependencies]
python = "^3.6"
coloredlogs = "^14.0"
GitPython = "^3.1.0"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
flake8 = "^3.7.9"
flake8-bugbear = "^20.1.4"
flake8-docstrings = "^1.5.0"
flake8-import-order = "^0.18.1"
mypy = "^0.770"
pylint = "^2.4.4"

[tool.poetry.scripts]
ansible-vault-filter-encrypt = 'ansible_vault_filter:encrypt'
ansible-vault-filter-decrypt = 'ansible_vault_filter:decrypt'

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
